requirements.txt
click==8.1.1
colorama==0.4.6
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.2.0
protobuf==4.21.12
Werkzeug==2.1.0
